Use template from root
